Better key frame trimming. Definitely set initial transform value.
Take advantage of the ReadOnlySpan for key frame trimming.
Ensure that initial values are set for Position and Anchor if key frame trimming removes all the frames. This same work needs to be done for other animatable properties, so I've added the ability to know whether the initial value needs to be set, but the code isn't yet handling it everywhere.