Add a new type to represent Animatables that have been trimmed to their context.
During translation, Animatables need to be considered in context. For example,
if an Animatable has its last key frame before the start time of the composition's inPoint,
its initial value is the value of its last key frame. To handle that, we add the TrimmedAnimatable
struct which is a wrapper around the contents of an Animatable.

Use the TrimmedAnimatable in the translator to ensure all IsAnimated checks are done
on the trimmed set of key frames, and that the InitialValue is correct in context.

Ensure values are not set on CompositionObjects if they have an animation. We always use
a key frame for the initial value if the object is animated so there is no need to
set the values directly on the object. Previously we did this optimally sometimes. Should
be much more consistent now.

Add a ContainerShapeOrVisual class so we only need one version of the TransformAndApplyTransform methods.

Make Animatable's constructors be explicit about animated-versus-non-animated versions. Previously
we would pass the InitialValue when constructing an animated Animatable but that was always the
value of the first keyframe. Rather than requiring the callers to get that right, add a constructor
that does the right thing for animated Animatables. This allows us to reason about Animatables
better because we now know by construction that the InitialValue is always the first keyframe value.

Make IsAnimated mean that an animatable has > 1 keyframes (was previously > 0). A single keyframe
does not represent an animation - you need at least 2.