Handle unreachable URLs as fallback cases.
Previously when an unreachable URL was set on the LottieVisualSource we'd return an animated visual that had no content.
However in order to trigger fallback behavior in the player, we have to return null. So now on any failure to load
we set a content factory that will return null, thus signaling the player to fallback.