Fixes NullReferenceException caused by codegen optimizer hole.
On rare cases the codegen optimizer that replaces methods calls with field accesses would pull from a field before it had been initialized. This was caused by the code generator generating nodes in a different order from what the ObjectGraph visited them. I've updated the ObjectGraph to visit the nodes in the same order as the code generator.