Give unique colors to nodes in DGML graph. (#28)
Making it easier to set colors  for DGML graph nodes, and giving each node different colors so that it's easier to tell what kind of geometry is attached to shape, and whether a node has any animations. This will make it easier to find opportunities for optimization.

Fix an assertion in the translator discovered during testing.
Make "percent" a first-class concept for properties in the translator to make maintenance easier.