Add ability to dump Lottie files as YAML.
The YAML is easier to read than XML. The YAML output exists to help debug issues with Lottie files. It's much easier to read the YAML than the raw JSON. Note that the dump is not complete - some properties may not show up. It's at least as good as the XAML however so checking it in, and we can add more properties as needed.