Adding LottieGenCorpus for generating output for testing. (#34)
* Adding LottieGenCorpus for generating output for testing.

LottieGenCorpus will create a directory with the output from running LottieGen over a corpus of Lottie files. The output can then be compared with a baseline run to detect regressions.

Add scripts to allow comparing the outputs of different versions of LottieGen

* Tweaking the tests scripts as a result of trying them a couple times.
Does a better job of finding the latest DLL, and gives more information during diffing.