Use D2D geometry groups.
This handles a specific case where there are multiple paths in the same shape layer.
Previously we just drew each path separately, which is not correct although it ends up being equivalent in many cases.
The symptom on the particular Lottie that was affected was that a box, made up of 2 paths, was filled with as solid color when it should have looked as if it was stroked.
The box was actually made up of 2 concentric boxes, and the fill was supposed to go between them. When drawn separately the outer box git filled and the inner box was not visible.
Now as a group, only the space between them is filled.