Do less InheritDoc work during build.
The InheritDocs task (that replaces <inheritdoc/> tags in doc comments with the comments from the
inherited member) was causing a lot of noise in local builds as it was running over every xml
file that had accumulated from various build configurations. We currently only care about
docs for one project, so I'm pointing the InheritDocs task at it so that it does less work
and is therefore less noisy.

I've also updated to the newest version of InheritDocs while I'm here, because, shiny.