Be more resilient to Lottie layers with invalid indices.
The Index of a Lottie layer is supposed to indicate its relative drawing order (layers are drawn from highest to lowest), as well as to allow it to be
referenced as the parent transform of another layer.

I've recently seen a Lottie with 2 layers having the same index. This was causing us to crash.

This change fixes that crash by assuming that when there are 2 or more layers with the same index that they are in the Lottie file in drawing order,
and that that index will never be used for transform inheritance. If they're not in drawing order then I don't know how any renderer would figure out how to draw them.
If the index is used to reference a parent then I don't know how any renderer would figure out which layer to return. So I think this behavior is reasonable
and definitely preferable to crashing.