Centralize list of warnings and errors. Generate artifacts from it.
We need to publish an HTML file that has more details about warnings and errors. But there are 3 different lists of issues (parsing, translation, and validation) and it would be difficult to keep everything in sync.
With this commit we move the data into Issues.xml and add scripts to generate the C# classes and the HTML.

Currently the generation has to be done manually, by:
1) Run GenerateIssuesClasses.cmd to update the C# classes. Requires texttransform.exe from the developer command prompt.
2) Run GenerateIssuesHtml.cmd to regenerate the html