Fix for #66 - cannot install on projects that have a MinVersion < RS5. (#67)
* Fix for #66 - cannot install on projects that have a MinVersion < RS5.

* Support back to 15063.

Sort of support 15063... no Lotties will actually render, but at least you should be able to build with a min version of 15063>

* Updating WinUI versions.

We were depending on a very old version of WinUI that had different versioning requirements. We need to be on this latest one to ensure we can be used in projects that have a minVer of 15063.

* Set the TargetFramework back to 16299 (RS3).

I was trying to set it to RS2 (15063) but we have taken a dependency on some .NET framework features that are only in 16299 and later. So rather than trying to remove those dependencies, we'll just support down to RS3 for now.