Override the default interpolation color space of ColorKeyFrameAnimation to align with After Effects. (#142)
After Effects (and the web) use RGB interpolation for color animations. Composition support HSL and RGB, but defaults to HSL. HSL is rarely useful if the animation goes through white or black because there are multiple representations of white and black so the interpolation is poorly specified.

With this change, Lottie-Windows always uses RGB interpolation.