Prerequisites for versioning. (#156)
* Work in progress towards versioning.
This adds a number of pieces required to generate version-adaptive code, but doesn't
yet hook those pieces up. This allows testing against a corpus to ensure nothing is
broken. A future commit will add version-adaptivity.

* Fix build break.

* CR style feedback.
I meant to fix this but must have forgotten. It was a byproduct of some experimentation.

* CR feedback.

* CR feedback.

* CR feedback.