Support gradient opacity stops. (#182)
* Support gradient opacity stops.

In LottieData we now model opacity and color stops on gradients as a single list of GradientStop which can either be an OpacityGradientStop or a ColorGradientStop.

During translation, we convert that list to a list of ColorGradientStop where the opacity information is carried in the Alpha channel.

During canonicalization, work around an issue with CompositionColorGradientStopCollections which require that a stop object does not appear twice in the same collection.

When translating gradients, use the maximum opacity in cases where we can't support opacity animation. Previously we used the initial opacity, which caused things to disappear if they started out with 0 opacity. Maximum opacity will allow the items to show up, which makes it easier to debug.

Updated the temporary cert used when building LottieViewer - the old one had expired.

* Update the samples cert - it had expired.