Adding a metadata facility to CompositionObject.
This adds a facility for arbitrary metadata to be added to CompositionObject and uses it for the IDescribable interface.
I also made the IDescribable implementation on CompositionObject explicitly implemented so that it doesn't pollute the API surface of CompositionObject.
And the IDescribable storage is now implemented by the metadata store, which should result in space savings.