Adding a property bindings language parser.
The parser is designed to allow designers to embed information about how colors should be hooked up to theme colors in the app or OS on which the Lottie runs.
Currently the parser is hooked up to parse property bindings from solid color fills, but it doesn't do anything with the result - this is just to get some coverage on it.
Eventually the property bindings will results in generated code that puts the given color in a CompositionPropertySet property and binds it appropriately so that clients can change the color at runtime.