Pass Lottie metadata (frame rate, etc) through translator to the code generators. (#197)
* Pass Lottie metadata (frame rate, etc) through translator to the code generators.
Add ability to convert GenericData object to code-genned Windows.Data.Json objects.

This change passes the data through but doesn't yet change the output of the code generators. That will come soon when we figure out what clients would like to have.

* Remove some dead code.

* Removing C#-8.0-ism from code used by UWP.

Switch expressions are new in C# 8.0. UWP requires C# 7.0 so the LottieVisualSource won't build with them.
Now that LottieVisualSource doesn't support codegen, we should refactor this so that codegen can use C# 8.0 (which is ok for LottieGen because it's .NET Core so can use C# 8.0).

* Remove another piece of C# 8.0 that snuck in. It breaks LottieVisualSource builds.