Make CompositionPath names start at 0. (#212)
CompositionPath factories get inlined during codegen unless they're shared. Because we were naming the factory nodes before determining whether the CompostionPath factories could be inlined, we ended up generating names for CompositionPaths that would never get a name. The result was that there would be gaps in the number of CompostionPath factories (e.g. CompostionPath_0, CompostionPath_4). With this change the CompostionPath numbers will be contiguous because we do the inlining before the naming.