Make expressions equatable.
This is needed for an unpcoming change which uses KeyFrame<Expressions.Color>. KeyFrame<T> requires that T is IEquatable (to support animation trimming).

Expressions are considered equivalent if their simplified text is exactly the same. It does not guarantee that the expression evaluates to the same value, but it does guarantee that if the expressions claim to be equal, they will evaluate to the same value.