Transfer object descriptions and names during tree reduction. (#229)
* Transfer object descriptions and names during tree reduction.

Tree reduction is an optimization on the Composition graph where we attempt to coalesce and remove redundant visuals and shapes. Previously this resulted in the loss of some descriptions and names on objects. Now we attempt to transfer that information out of the objects that are being removed into the objects that are taking over their roles.