Make the YAML output even more human readable by including comments. (#228)
This change makes the YAML output work better with YAML editors (in particular VS Code) when they collapse sequences. Previously a collaped item in a sequence would be invisible until it was uncollapsed. Now, you'll see a comment for the item.

To take advantage of this new behavior, the Lottie yaml serializer now adds a comment containing the name of each object.

And we no longer display the MatchName field (as it's rarely used and is meaningless to us).

The result is that Lottie YAML files are much easier to navigate, as you can collapse parts of them to get an overview without losing information about what the collapsed items are.