Use nameof to keep Yaml field names in sync with Lottie data property names.
Some of the Yaml field names were wrong. This change gets rid of the hand-written strings and replaces them with nameof(...) so that the names stay in sync with LottieData.
And did some small cleanup and refactoring while I was there.

This does change the Yaml output from what it was - that is intentional.