Simplify node naming. (#234)
* Simplify node naming.

Previously we special-cased the naming of the root node and the root progress animations during codegen, but did some other special naming during translation. With this change we do all of the special naming at codegen time.

This change also includes better handling of name collisions in the node namer. If 2 or more nodes get the same name, we have to append a counter value to the name to differentiate them. Previously we appended a counter even if there was only 1 node, in order to prevent C++ from getting confused when a method name was the same as a type name. Now we keep track of whether a name is also a type name, and can therefore avoid appending the counter more often. This makes for much nicer naming (i.e. they don't have a _0 suffix) for most singleton objects.