Add ability to automatically bind to the colors in a Lottie's palette.
When enabled with -GenerateColorBindings, LottieGen will create bindings
for each non-animated color it finds in a fill or stroke. This will
allow the colors to be changed at runtime.

This is particularly handy for animations that have a single color such
as those found at https://www.lottieflow.com.

The feature is less flexible than theme property bindings because:
a) it only supports colors
b) it doesn't give the designer much control over which parts of the
animation are bindable.
c) it doesn't allow the designer to choose the names used for the
bindings.

The binding name is the color, so a black color will generate a property
called "ColorBlack".

For single-color Lottie files we could be clever and name the property
"Foreground". I'm not doing that right now because we don't yet know
how the feature will be consumed.