Fix rare case name collision in generated code. (#314)
Very rarely the generated code contained name collisions as a result of a name having the same form as a name with a counter suffix, e.g. Rectangle_15 looks like a name with a counter suffix but it might just be a 15x15 rectangle.
We now check for collisions when adding a counter suffix.