Eliminate the ReadOnlySpan from Sequence. (#313)
This has been bugging me for ages. It was an attempt to switch the code over to using Span, but in the end it didn't make thing better, and required extra indirections to get at the items in the sequence.
While I'm here I'm adding some other nice touches, including avoiding copying of arrays when create a sequence, and allowing the creation of a sequence from a single item.