Use clearer names for automatically generated color bindings. (#321)
When LottieGen generates color bindings automatically, it creates WinRT properties for each distinct color in the Lottie. The names of these properties were using the descriptive color names which are great for trying to understand code, but they're unnecessarily complex for property names. For example, naming a color AlmostTurquoise_FF123456 is including information about the color that the user doesn't care about. The user simply wants to identify which property is which. So now we just output "Color_XXXXXX" where XXXXXX is a 6 digit hex represenation of the color, ignoring alpha.

This change also fixes a bug I noticed during testing where the enumerators for sliced lists were never created.