Invalidated IAnimatedVisual on image loading completed, even if it's not successful. (#339)
The code generated for loading an IDynamicAnimatedVisualSource previously only invalidated when all of the images loaded successfully. This means that any failure to load an image would result in the IAnimatedVisual never loading.
With this change, completion of all image loads, even if one of them fails, will invalidate the IDynamicAnimatedVisualSource.
This will allow a Lottie to show up even if one of its images can't load.