Show animator counts in the object status table in generated code. (#343)
We were showing the number of animation objects previously, but that doesn't really tell you how expensive animations are in DWM because the one animation might be started multiple times. Animator counts are a better representation of cost.

Here's an example of the extra info. Note the animators, and parameters counts.

```// |       Object stats       | UAP v12 count | UAP v8 count |
// |__________________________|_______________|______________|
// | All CompositionObjects   |           567 |          567 |
// |--------------------------+---------------+--------------|
// | Expression animators     |            51 |           51 |
// | KeyFrame animators       |            50 |           50 |
// | Reference parameters     |            51 |           51 |
// |--------------------------+---------------+--------------|
// | Animated brushes         |             1 |            1 |
```