Change the default for IsAnimatedVisualSourceDynamic.
This only affects Lotties with non-embedded images. When IsAnimatedVisualSourceDynamic is true the load of the Lottie is not considered complete until all the images have loaded. This guarantees that you never see a Lottie that is missing its images.
When IsAnimatedVisualSourceDynamic is false, the Lottie will load synchronously and the image will pop in when they are finally loaded.
Lotties with non-embedded images are quite rare, and one dev that tried to use them found them to be difficult because they had to wait for the Lottie to be loaded before playing. But they really didn't care about the images taking too long (they were on disk so they should show up really quickly). If we had made IsAnimatedVisualSourceDynamic default to false they wouldn't have had to deal with waiting for the loading and would never have noticed any issue.
So, assuming that most use case are similar (i.e. the images are on disk and there is no need to wait for the images to load), I'm changing the default value false. You can always turn on IsAnimatedVisualSourceDynamic if you need it, but I'm pretty sure hardly anyone ever will.