Enable nullable reference types where possible.
This enables C# 8 nullable reference types in the parts of the codebase where we always build with C# 8.
Most of the codebase is built using into winrt components and apps and therefore can't use C# 8 yet. That leaves the LottieGen command line interface and the code generator as the only parts that can use C# 8.