Implement DropShadow effect.
This is currently turned off, however we do everything except adding the shadow to the LayerVisual so that the code is being exercised.
I have not output any issues yet for the cases we don't handle because we don't know how long before LayerVisual is fixed, and the set of issues might need changing when LayerVisual is finally working.